Turns llm stream into a reusable functions
Refactor sendMessage and editMessage to use a common function called handleLLMStream
